requirements.txt
folium
geocube
geopandas
nifc_wildfires
numpy
pandas
pyproj
rasterio
shapely
streamlit
streamlit_folium
colorama

python-multipart==0.0.6
fastapi==0.97.0
opencv-python==4.7.0.72
pytest-asyncio==0.21.0
httpx==0.24.1
matplotlib==3.5.3
uvicorn==0.22.0
Pillow==9.5.0
branca==0.3.1
earthengine-api==0.1.357
google.cloud

# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.12.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.12.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.12.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
